fix validation failure with variables in multiple operations
When a document with multiple operations has variables with different
types, a validation error would occur, because the validation checked
that the type of the variable matched the type of *all* places it's
used. In multi-operation documents, variables only apply to the selected
operation.

Don't check argument types that are variables when in unselected
operations.