Preserve underscore when renaming field
Fixes https://github.com/async-graphql/async-graphql/issues/1698

I had to change a bunch of places that relied on variable names with
underscores not being preserved.

So technically I think this is a breaking change - but I also think it
was a bug to not preserve the underscore.

What do you think?